+++ pkg/wharfapi/build.go
@@ -6,6 +6,8 @@ import (
 	"gopkg.in/guregu/null.v4"
 )
 
+// Build is the metadata about a code execution for a certain project inside
+// Wharf.
 type Build struct {
 	BuildID     uint         `json:"buildId"`
 	StatusID    BuildStatus  `json:"statusId"`
@@ -20,8 +22,10 @@ type Build struct {
 	IsInvalid   bool         `json:"isInvalid"`
 }
 
+// BuildParam is an input parameter provided by the user or service that started
+// the build.
 type BuildParam struct {
-	BuildID      uint   `json:"buildId"`
-	Name         string `json:"name"`
-	Value        string `json:"value"`
+	BuildID uint   `json:"buildId"`
+	Name    string `json:"name"`
+	Value   string `json:"value"`
 }