+++ pkg/wharfapi/buildlog.go
@@ -8,11 +8,15 @@ import (
 	"time"
 )
 
+// BuildLog is a single log line from a given build. The Timestamp refers to
+// when the logged line was printed inside the build.
 type BuildLog struct {
 	Message   string    `json:"message"`
 	Timestamp time.Time `json:"timestamp"`
 }
 
+// PostLog adds a new log to a build by invoking the HTTP request:
+// 	POST /api/build/{buildID}/log
 func (c Client) PostLog(buildID uint, buildLog BuildLog) error {
 	body, err := json.Marshal(buildLog)
 	if err != nil {
@@ -28,6 +32,8 @@ func (c Client) PostLog(buildID uint, buildLog BuildLog) error {
 	return nil
 }
 
+// PutStatus updates a build by invoking the HTTP request:
+// 	PUT /api/build/{buildID}
 func (c Client) PutStatus(buildID uint, statusID BuildStatus) (Build, error) {
 	uri := fmt.Sprintf("%s/api/build/%d?status=%s", c.APIURL, buildID, url.QueryEscape(statusID.String()))