+++ pkg/wharfapi/buildstatus.go
@@ -2,12 +2,26 @@ package wharfapi
 
 import "strconv"
 
+// BuildStatus is the state of a build.
+//
+// The flow of the build status goes like this:
+//
+//            Scheduling
+//                |
+//                V
+//             Running
+//               / \
+//  Completed <-/   \-> Failed
 type BuildStatus int
 
 const (
+	// BuildScheduling means that the build has not started execution yet.
 	BuildScheduling = BuildStatus(iota)
+	// BuildRunning means that the build is currently executing code.
 	BuildRunning
+	// BuildCompleted means that the build ran successfully to completetion.
 	BuildCompleted
+	// BuildFailed means that the build ran unsuccessfully.
 	BuildFailed
 )