+++ CHANGELOG.md
@@ -16,14 +16,15 @@ This project tries to follow [SemVer 2.0.0](https://semver.org/).
 
 - Added documentation to all exported methods. (#11)
 
-- Added dependency on `github.com/iver-wharf/wharf-core`. (#12, #13)
+- Added dependency on `github.com/iver-wharf/wharf-core` v1.1.0.
+  (#12, #13, #16)
 
 - Changed all logging via `fmt.Print` and sirupsen/logrus to instead use the new
  `github.com/iver-wharf/wharf-core/pkg/logger`. (#13)
 
 - Changed to use `problem.Response` from `wharf-core` instead of the
   `wharfapi.Problem` that was added in v1.3.0/#4. (#12, #14)
-  
+
 - Removed `wharfapi.ProblemError` as the `problem.Response` type from the
   `wharf-core` module conforms to the `error` interface. (#14)